Ajustes fluxo reservation

+++ src/main/java/br/com/livelo/orderflight/domain/dtos/pricing/response/PricingCalculatePrice.java
@@ -12,10 +12,10 @@ import java.math.BigDecimal;
 public class PricingCalculatePrice {
     private String priceListId;
     private BigDecimal amount;
-    private Integer pointsAmount;
+    private BigDecimal pointsAmount;
     private PricingCalculateFlight flight;
     private PricingCalculateTaxes taxes;
-    private Integer accrualPoints;
+    private BigDecimal accrualPoints;
     private PricingCalculatePricesDescription pricesDescription;
     private Float multiplier;
     private Double multiplierAccrual;
