+++ src/main/java/br/com/livelo/orderflight/domain/entity/OrderPriceEntity.java
@@ -21,7 +21,7 @@ public class OrderPriceEntity extends BaseEntity {
     @SequenceGenerator(name = "ORDERS_PRICE_SEQ", sequenceName = "ORDERS_PRICE_SEQ", allocationSize = 1)
     @Id
     private Long id;
-    private Double accrualPoints; //RECUPERAR DA PRECIFICAÇÃO
+    private BigDecimal accrualPoints; //RECUPERAR DA PRECIFICAÇÃO
     private BigDecimal amount; //RECUPERAR DA PRECIFICAÇÃO
     private BigDecimal pointsAmount; //RECUPERAR DA PRECIFICAÇÃO
     private BigDecimal partnerAmount;
